Input field displays. CSS edits. Doesn't do anything as far as I know.
Continuing to work on adding type filter. The input field now displays
as it should.

Still having trouble querying the category type. As a test, I tried
creating categoryTest in app\models\question.js (and adding a field to
the questions on the WP site) as a data type I could base a beta filter
off of, however I could not get it to display in the
components\question-list.hbs template, so I don't think it's being
called correctly either. Will continue to troubleshoot.

Minor CSS edits to input field and question divs.