Allows for string-type answers. Succesful test of additional category.
Answer attribute of question model no longer needs to be a number.

An additional category (Writing) was tested and successfully worked with
the new filter.

Still: aforementioned array issue and the fact that only category ID#
can be input. Array-based dictionary between category ID# and name might
be useful for input.