Filter working again (I think)
Fixed a strange bug where the filter stopped working suddenly. Odd,
cause the "working" version earlier had errors in questions.js (a period
in the place of a semicolon [line 7]) and yet still worked for a while
last night. I could not manage to get it working again, even after
restoring files to the earlier state. Who knows?

Anyways, it is working again on my end, although my confidence is shaken
as to whether it will stay that way. Will continue to test to make sure.