Image field added; Selected answer stays highlighted
Added an image field that expands and shrinks based on a click. Could be
useful for math figures. Right now, the image is static, is erroneously
used for all questions, and is not using the WP-API. I will work to fix
this soon.

I also added the feature where a selected answer stays highlighted. In
order to do this, I installed "ember-truth-helpers" which allows the
{{if}} handlebars helper to do basic javascript logic comparisons. Then
I created a variable "isSelected" that is set to the clicked answer and
determines whether or not the answer option is highlighted.

See (https://github.com/jmurphyau/ember-truth-helpers) for documentation
on ember-truth-helpers.

I also made some minor CSS changes to get these two features working.